app/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.2
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
django-jazzmin==2.6.0
redis==5.0.1
djangorestframework-api-key==3.0.0
gunicorn==21.2.0
drf-yasg==1.21.7
django-redis==5.4.0
